requirements.txt
pyyaml>=6.0.1            # For reading/writing YAML files
icalendar>=5.0.11        # For generating iCal files
beautifulsoup4>=4.12.3   # For cleaning HTML descriptions
requests>=2.31.0         # For Meetup API calls
python-dateutil>=2.8.2   # For date manipulation
feedparser>=6.0.11       # For RSS feeds parsing (if needed)
pytz>=2024.1            # For timezone handling
markdown2>=2.4.12       # For README formatting
PyGithub==2.1.1         # For GitHub API interactions

